+++ hapi-fees/src/main/java/com/hedera/services/usage/file/FileOpsUsage.java
@@ -9,9 +9,9 @@ package com.hedera.services.usage.file;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -70,9 +70,9 @@ public class FileOpsUsage {
 
 		var estimate = txnEstimateFactory.get(sigUsage, fileCreation, ESTIMATOR_UTILS);
 		/* Variable bytes plus a long for expiration time */
-		estimate.addBpt(customBytes + LONG_SIZE);
+		estimate.addBpt((long) customBytes + LONG_SIZE);
 		estimate.addSbs((bytesInBaseRepr() + customBytes) * lifetime);
-		estimate.addNetworkRbs(BASIC_ENTITY_ID_SIZE * USAGE_PROPERTIES.legacyReceiptStorageSecs());
+		estimate.addNetworkRbs((long) BASIC_ENTITY_ID_SIZE * USAGE_PROPERTIES.legacyReceiptStorageSecs());
 
 		return estimate.get();
 	}
